Update book and car class examples

+++ OOPs/class_and_objects/class_of_Books.cpp
@@ -5,7 +5,7 @@ class book{
 public:
     string title;
     string author;
-    double pages;
+    int pages;
     int price;
 
     void display(){
